requirements.txt
annotated-types==0.7.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
numpy==2.0.0
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyairtable==2.3.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
